+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 * Updated mimalloc lib for static builds to latest v3.5.3.
 * Updated S3 to latest AWS SDK CPP v1.11.895.
 * Added test suite to repo based on "prove". `tests/run-tests.sh -a` can be used to run all tests supported by the current executable in `bin/`. (Tests for not included build features will automatically be skipped.) Details available at [`tests/README.md`](tests/README.md).
+* Replaced the AWS SDK logger (`--s3log`) by an implementation that does not do application-level buffering, so that all log lines are guaranteed to appear in the log file after a service mode benchmark phase has ended.
 
 ### Fixes
 * The AWS SDK CPP for S3 differs in retry behavior from the public documentation [here](https://docs.aws.amazon.com/sdkref/latest/guide/feature-retry-behavior.html), so that e.g. http response code 429 (Too many requests) does not always lead to the intended retries. Thus, we now manually mark such reponse codes as retryable for the AWS SDK CPP in elbencho's custom retry strategy class.